fix: history component
