+++ src/shared/types/icon.types.ts
@@ -1,5 +1,4 @@
 export interface IconProps {
     height?: number
     width?: number
-    onClick?: () => void
 }
